requirements.txt
streamlit
openai-whisper
streamlit-mic-recorder
torch
ffmpeg-python